+++ models/user_model.py
@@ -1,0 +1,9 @@
+from beanie import Document
+from pydantic import EmailStr
+
+class User(Document):
+    email: EmailStr
+    password: str
+
+    class Settings:
+        name = "users"
